backend/requirements.txt
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.1.2
boto3==1.42.44
botocore==1.42.44
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
jmespath==1.1.0
Mako==1.3.10
MarkupSafe==3.0.3
moncashify==1.1.0
passlib==1.7.4
paypalrestsdk==1.13.3
pillow==12.1.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-magic-bin==0.4.14
python-multipart==0.0.21
python-slugify==8.0.4
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
stripe==14.4.0
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
